Added path_to_ai_agent in config to specify which agent to consider to create bots when a client disconnects
